AssayResultsForSamplesButton: move to @labkey/components/entities
